fix: render only 4 cards
