+++ astro-frontend/src/components/pastoral-council-card.astro
@@ -1,14 +1,14 @@
 ---
+import { Image } from "astro:assets";
 import { Locale } from "~/enums/locale";
 import { cn } from "~/helpers/cn.ts";
 import { getLangFromUrl } from "~/i18n/utils";
 import NetworkImage from "./ui/network-image.astro";
-import { getPlaceholderImage } from "~/helpers/get-placeholder-image";
+import profileImage from "~/assets/parish-priests/blank.jpeg";
 
 interface Props {
 	src: string;
-	baseUrl?: string;
-	width: number;
+  width: number;
 	height: number;
 	alt?: string;
 	class?: string;
@@ -27,20 +27,17 @@ const { src, width, height, class: className, name, position} = Astro.props;
   )}
 >
   <div class="card-image flex-none">
-    <NetworkImage
-    src={
-      src
-        ? new URL(
-            src,
-            import.meta.env.STRAPI_URL
-          ).toString()
-        : getPlaceholderImage()
-    }
-    width={width}
-    height={height}
-    alt={`${position} of Our Lady of Pompei Church`}
-    class="h-[300px] w-full object-cover"
-    />
+    {src ? (
+      <NetworkImage
+        src={src}
+        width={width}
+        height={height}
+        alt={`${position} of Our Lady of Pompei Church`}
+        class="h-[300px] w-full object-cover"
+      />
+    ) : (
+      <Image src={profileImage} alt="" class="h-[300px] w-full object-cover" width={width} height={height} />
+    )}
   </div>
 
   <div
