+++ astro-frontend/src/constants/message.ts
@@ -5,5 +5,6 @@ export class Message {
   public static readonly COUNCIL_MEMBERS_NOT_FOUND = 'No Council Members Found!';
   public static readonly MAGAZINES_NOT_FOUND = 'No Magazines available!';
   public static readonly POPES_INTENTION_NOT_FOUND = 'Popes Intention Not found!';
-  public static readonly CORE_COMMITTEE_NOT_FOUND = ' No Core Committee Found!';
+  public static readonly CORE_COMMITTEE_NOT_FOUND = 'No Core Committee Found!';
+  public static readonly FINANCE_COMMITTEE_NOT_FOUND = 'No Finance Committee Found!';
 }
