+++ astro-frontend/src/pages/parish-pastoral-council.astro
@@ -3,7 +3,8 @@ import Layout from "~/layouts/layout.astro";
 import { getLangFromUrl, useTranslations } from "~/i18n/utils";
 import ParishPastoralCouncilList from "~/components/sections/parish-pastoral-council/parish-pastoral-council-list.astro";
 import type { ParishPastoralCouncil } from "~/models/parish-pastoral-council";
-import { listPastoralCouncilData, getPastoralCouncilImage } from "~/services/parish-pastoral-council-data.ts";
+import { listPastoralCouncilData } from "~/services/parish-pastoral-council/parish-pastoral-council-data.ts";
+import { getPastoralCouncilImage } from "~/services/parish-pastoral-council/parish-pastoral-council-image.ts";
 import { Message } from "~/constants/message.ts";
 import { Locale } from "~/enums/locale";
