+++ astro-frontend/src/services/parish-pastoral-council/parish-pastoral-council-core-committee-data.ts
@@ -12,15 +12,19 @@ async function listPastoralCoreCommitteeData(args?: {
   sortBy?: string;
   locale?: Locale;
 }): Promise<PastoralCoreCommittee[]> {
-
-  const { page = 1, pageSize = 25, sortBy = "sNo:asc", locale = Locale.EN } = args ?? {};
+  const {
+    page = 1,
+    pageSize = 25,
+    sortBy = "sNo:asc",
+    locale = Locale.EN,
+  } = args ?? {};
 
   const queryParams = new URLSearchParams({
-    'populate[0]': 'image',
-    'pagination[page]': String(page),
-    'pagination[pageSize]': String(pageSize),
-    'sort[0]': sortBy,
-    'locale': locale,
+    "populate[0]": "image",
+    "pagination[page]": String(page),
+    "pagination[pageSize]": String(pageSize),
+    "sort[0]": sortBy,
+    locale: locale,
   });
 
   const data = await strapiFetch<PastoralCoreCommitteeData>({
@@ -28,11 +32,7 @@ async function listPastoralCoreCommitteeData(args?: {
     queryParams,
   });
 
-  if (!data?.data) {
-    return null;
-  }
-
-  return data.data;
+  return data?.data ?? [];
 
 }
