+++ astro-frontend/src/services/parish-pastoral-council/parish-pastoral-council-data.ts
@@ -4,8 +4,6 @@ import { strapiFetch } from "~/helpers/strapi-fetch";
 import type {
   ParishPastoralCouncil,
   ParishPastoralCouncilData,
-  ParishPastoralCouncilImage,
-  ParishPastoralCouncilImageData,
 } from "~/models/parish-pastoral-council";
 
 /**
@@ -23,13 +21,18 @@ async function listPastoralCouncilData(args?: {
   sortBy?: string;
   locale?: Locale;
 }): Promise<ParishPastoralCouncil[]> {
-  const { page = 1, pageSize = 25, sortBy = "sNo:asc", locale = Locale.EN } = args ?? {};
+  const {
+    page = 1,
+    pageSize = 25,
+    sortBy = "sNo:asc",
+    locale = Locale.EN,
+  } = args ?? {};
 
   const queryParams = new URLSearchParams({
-    'pagination[page]': String(page),
-    'pagination[pageSize]': String(pageSize),
-    'sort[0]': sortBy,
-    'locale': locale,
+    "pagination[page]": String(page),
+    "pagination[pageSize]": String(pageSize),
+    "sort[0]": sortBy,
+    locale: locale,
   });
 
   const data = await strapiFetch<ParishPastoralCouncilData>({
@@ -38,24 +41,7 @@ async function listPastoralCouncilData(args?: {
   });
 
   return data?.data ?? [];
+
 }
 
-async function getPastoralCouncilImage(): Promise<ParishPastoralCouncilImage | null> {
-	const queryParams = new URLSearchParams({
-		'populate[0]': 'pastoralCouncilImage',
-	});
-
-	const data = await strapiFetch<ParishPastoralCouncilImageData>({
-		endpoint: ROUTES.PARISH_PASTORAL_COUNCIL_IMAGE,
-		queryParams,
-	});
-
-
-  if (!data?.data) {
-    return null;
-  }
-
-  return data.data;
-}
-
-export { listPastoralCouncilData, getPastoralCouncilImage};
+export { listPastoralCouncilData };
