+++ strapi-backend/src/api/parish-pastoral-council-core-committee/content-types/parish-pastoral-council-core-committee/schema.json
@@ -19,7 +19,7 @@
     "image": {
       "type": "media",
       "multiple": false,
-      "required": true,
+      "required": false,
       "allowedTypes": [
         "images"
       ],
